+++ storage/backup/backup-service/src/handlers/mod.rs
@@ -6,15 +6,12 @@ mod bytes_sender;
 mod utils;
 
 use crate::handlers::utils::{
-    handle_rejection, reply_with_async_channel_writer, reply_with_bcs_bytes,
-    reply_with_bytes_sender, send_size_prefixed_bcs_bytes, size_prefixed_bcs_bytes, unwrap_or_500,
+    handle_rejection, reply_with_bcs_bytes, reply_with_bytes_sender, unwrap_or_500,
     LATENCY_HISTOGRAM,
 };
 use aptos_crypto::hash::HashValue;
 use aptos_db::backup::backup_handler::BackupHandler;
-use aptos_storage_interface::AptosDbError;
 use aptos_types::transaction::Version;
-use hyper::Body;
 use warp::{filters::BoxedFilter, reply::Reply, Filter};
 
 static DB_STATE: &str = "db_state";
@@ -49,26 +46,10 @@ pub(crate) fn get_routes(backup_handler: BackupHandler) -> BoxedFilter<(impl Rep
     let bh = backup_handler.clone();
     let state_snapshot = warp::path!(Version)
         .map(move |version| {
-            let (mut sender, stream) = bytes_sender::BytesSender::new();
-
-            // spawn and forget, error will propagate through the stream
-            let bh = bh.clone();
-            let _join_handle = tokio::task::spawn_blocking(move || {
-                if let Err(err) = {
-                    for res in bh.get_account_iter(version)? {
-                        let record = res?;
-                        let bytes = size_prefixed_bcs_bytes(&record)?;
-                        sender.send_bytes(bytes)?;
-                    }
-                    Ok(())
-                } {
-                    sender.abort::<AptosDbError>(err)
-                } else {
-                    sender.finish()
-                }
-            });
-
-            Box::new(warp::reply::Response::new(Body::wrap_stream(stream)))
+            reply_with_bytes_sender(&bh, STATE_SNAPSHOT, move |bh, sender| {
+                bh.get_account_iter(version)?
+                    .try_for_each(|record_res| sender.send_size_prefixed_bcs_bytes(record_res?))
+            })
         })
         .recover(handle_rejection);
 
@@ -96,14 +77,9 @@ pub(crate) fn get_routes(backup_handler: BackupHandler) -> BoxedFilter<(impl Rep
     let bh = backup_handler.clone();
     let transactions = warp::path!(Version / usize)
         .map(move |start_version, num_transactions| {
-            // use async move block to group `bh` and the iterator into the same lifetime, since the
-            // latter references the former.
-            reply_with_async_channel_writer(&bh, TRANSACTIONS, |bh, sender| async move {
-                send_size_prefixed_bcs_bytes(
-                    bh.get_transaction_iter(start_version, num_transactions),
-                    sender,
-                )
-                .await
+            reply_with_bytes_sender(&bh, TRANSACTIONS, move |bh, sender| {
+                bh.get_transaction_iter(start_version, num_transactions)?
+                    .try_for_each(|record_res| sender.send_size_prefixed_bcs_bytes(record_res?))
             })
         })
         .recover(handle_rejection);
